+++ CHANGELOG.md
@@ -7,3 +7,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 #### Added
 - Initial support for port scanner added
+- Send alert to slack if any panic alert is faced
+- Send start scan and end scan updates to dast api server